moved particles to App, updated gh pages
