+++ src/Portfolio.js
@@ -1,9 +1,6 @@
 import "./Portfolio.css";
 import logo from "./assets/img/portfolioSelfie2.jpg";
 import { CustomLink } from "./components/CustomLink";
-import Particles, { initParticlesEngine } from "@tsparticles/react";
-import { loadSlim } from "@tsparticles/slim"; // if you are going to use `loadSlim`, install the "@tsparticles/slim" package too.
-import { useState, useEffect, useMemo } from "react";
 import { Section } from "./components/Section";
 import { SkillItem } from "./components/SkillItem";
 
@@ -17,106 +14,8 @@ const linksLoop = (links) => {
 };
 
 const Portfolio = (props) => {
-  const [width] = useState(window.innerWidth);
-  const isMobile = width <= 768;
-  const [loadedParticles, setLoadedParticles] = useState(false);
-
-  // this should be run only once per application lifetime
-  useEffect(() => {
-    initParticlesEngine(async (engine) => {
-      // you can initiate the tsParticles instance (engine) here, adding custom shapes or presets
-      // this loads the tsparticles package bundle, it's the easiest method for getting everything ready
-      // starting from v2 you can add only the features you need reducing the bundle size
-      await loadSlim(engine);
-    }).then(() => {
-      setLoadedParticles(true);
-      console.log("Loaded Particles Engine");
-    });
-  }, []);
-
-  // Options for Particles
-  const options = useMemo(
-    () => ({
-      // background: {
-      //   color: {
-      //     value: "#0d47a1",
-      //   },
-      // },
-      // fpsLimit: isMobile ? 60 : 120,
-      fpsLimit: 120,
-      detectRetina: true,
-      fullScreen: {
-        enable: true,
-        zIndex: 0,
-      },
-      pauseOnBlue: true,
-      pauseOnOutsideViewport: true,
-      interactivity: {
-        events: {
-          onClick: {
-            enable: true,
-            mode: "push",
-          },
-          onHover: {
-            enable: true,
-            mode: "repulse",
-          },
-        },
-        modes: {
-          push: {
-            quantity: 4,
-          },
-          repulse: {
-            distance: 150,
-            duration: 0.4,
-          },
-        },
-      },
-
-      particles: {
-        color: {
-          value: "#ffffff",
-        },
-        links: {
-          color: "#000",
-          distance: 150,
-          enable: true,
-          opacity: 1,
-          width: 2,
-        },
-        move: {
-          direction: "none",
-          enable: true,
-          outModes: {
-            default: "bounce",
-          },
-          random: false,
-          speed: 4,
-          straight: false,
-        },
-        number: {
-          density: {
-            enable: true,
-          },
-          value: 80,
-        },
-        shape: {
-          type: "circle",
-        },
-        size: {
-          value: { min: 1, max: 4 },
-        },
-      },
-    }),
-    []
-  );
-
   return (
     <>
-      <Particles
-        options={options}
-        className={loadedParticles ? "fade-in" : ""}
-      />
       <div className="intro floating">
         <img className="logo" src={logo} alt="This is my selfie" />
